Feature/api python3 12 (#259)
* Updated Dockerfile to use Python 3.12

* Update python requirements to new package versions

* Fixed pydantic BaseSettings breaking change

* Renamed schema_extra to json_schema_extra

* data_management: Changed config class overload to ConfigDict

* data_management: Replaced deprecated utcnow() with now(UTC)

* data_management: Replaced deprecated dict() method with model_dump()

* data_management: Fixed typing issues

- Made None initialized Members optional
- Annotated obscure data types

* data_management: deleted unused variables to remove warnings

* tests: fixed case test

* tests: fixed diagnosis test

* data_management: fixed typing issues in timeseries_data

* security: Changed deprecated root_validator to model_validator

* tests: Changed deprecated utcnow() to now(UTC)

* data_management: Made diagnosis bucket optional

* data_management: Introduced Self type to case

* data_management: shortend index annotations

* data_management: fixed typing issues in case.py

* diagnostics_management: Replaced Union syntax and made kg_url optional

* diagnostics_management: Made _celery in tasks.py optional

* security: Made None initialized variables optional

* storage: Removed unneeded data parameter

* routers: fixed typing issues in shared.py

* routers: Added return types to minio.py

* storage: Added missing data argument to put_data

* tests: Added asyncio decorators to diagnostics_management

* tests: Changed deprecated .utcnow() to .now(UTC)

* tests: Fixed httpx client deprecation warning

* tests: Added TODO comments

* tests: Added new asyncio decorator to routers

* tests: Added new asyncio decorator to data_management

* tests: Added new asyncio decorator to key auth test.

* tests: fixed api key auth HPPTException not getting unpacked

* tests: fixed keycloak test

* tests: Changed deprecated decorators to contextmanager in workshop router

* tests: fixed typing issues ind workshop router test

* tests: changed deprecated dict() with model_dump() in workshop router test

* tests: Fixed typing issues in data_management

* tests: Moved diagnostics router tests to loop_scope "module"

* api: Added typing annotations and exceptions to diagnostics router

* api: added typing annotations and renamend variables in shared router

* api: Added typing annotations to workshop router

* api: Fixed typing issues and added exceptions in workshop router

* tests: Added asserts to diagnostics router test

* Added pyproject.toml to configure pytest

* 255 filereader samplingrate issue (#256)

* Added rounding of duration and sampling_rate for picoscope csv

sampling_rate and duration will be rounded to nearest int.

* Added rounding of duration and sampling_rate for picoscope mat

Rounding duration and sampling_rate to nearest int.

* Added more type annotations for picoscope csv filereader

* Added more type annotations for picoscope mat filereader

* Fixed flake8 issues

* tests: (Hack) Added hack to timeseries_data test in data_management

Motor and pytest_asyncio don't play nice here running in different
event_loops (motor seems to generate their own depending on the context
motor functions are called).
To fix this the signal_bucket has to be created in the async function
instead of the fixture in conftest.
This solution is considered a hack.

* tests: renaming test_app to app since the test_ prefix triggers pytest

* Added warning filter for FastAPI wrong override of app variable

The FastAPI TestClient set it's app Member. This is considered
deprecated by httpx since transport is to be used. TestClient does use
transport but still overrides app which results in the warning.
Maybe TestClient should be removed in favor of httpx Client and
AsyncClient.

* api: Fixed exclude schema_version in Case

* api: fixed exclude id in Vehicle

* api: fixed issue in minio_storage

* tests: fixed typing issues in case data_management

* tests: Using BaseSingalStore instead of None

signal_store is not allowed to be None therefore BaseSingalStore is
being used instead since it throws exceptions on member function calls.

* tests: Ignore pyright unable to resolve bucket.collection

* api: Ignoring wierd mismatch between BinaryIO and IO[bytes]

* api: fixed flake8 issues

* tests: fixed flake8 issues

* Revert "tests: (Hack) Added hack to timeseries_data test in data_management"

This reverts commit b34604a6916f9afa54a29c573755ce5438a1d879.

* tests: Fixed fixtures and tests being in different loop scopes

* api: Ignore warning about Settings() without arguments

* tests: Removing ** unpack syntax since it cannot be typed checked

* Added missing space in Dockerfile

* Reformated code using Ruff

* Updated docker test workflow to use python 3.12

* Trying new ubuntu image instead of custom python install

* Install python packages in venv instead of system-wide

Installing system-wide python packages over pip is no longer allowed. A
virtual environment has to be used.

* Splitting up pytest/venv into setup venv, install depends, run pytest

* venv activation does not persist between steps. Activating if needed.

* Revert "Reformated code using Ruff"

This reverts commit 93b1f7de7fc6841826784cd7b3d453848a6cfe86.

* data_management: Added default value to customer_id

* data_management: Fixed typing and deprectation issues in customer.py

* routers: Updated deprecated .dict() to .model_dump()

* router: renamed customer_id to customer_oid

Using customer_id overrides variable of type str with ObjectID.
Thereforce the customer_oid variable is introduced to avoid override.

* tests: Fixed small issues in data_management tests

* tests: Changed deprecated utcnow() to now(UTC) in routers

* tests: Added missing asserts to customer router tests

* tests: Changed base_url in customer router tests to testserver

If testserver is ommited httpx will add a port notation to the Address.
The Adress will start with "http://:None" which will be reflected by
pagination and results in errors on calling the next entries.
This is probably an issue of httpx.

* Reverting the usage of Sequence back to List [skip ci]

* router: Updated HTTPException status code and detail in diagnostics

* router: Updated HTTPException status code and detail in workshop

* Updated api requirements to fix pymongo dependecy error in motor

- Updated motor to version 3.6.0
- Set pymongo to latest known working state (version 4.9.1)

+++ api/api/data_management/case.py
@@ -1,11 +1,28 @@
-from datetime import datetime
+from datetime import datetime, UTC
 from enum import Enum
-from typing import List, Union, Optional
+from typing import (
+    Annotated,
+    List,
+    Optional,
+    Tuple,
+    Any,
+    Self
+)
 
 from beanie import (
-    Document, Indexed, before_event, Insert, Delete, PydanticObjectId
+    Document,
+    Indexed,
+    before_event,
+    Insert,
+    Delete,
+    PydanticObjectId
 )
-from pydantic import BaseModel, Field, NonNegativeInt
+from pydantic import (
+    BaseModel,
+    Field,
+    NonNegativeInt,
+    ConfigDict
+)
 
 from .diagnosis import Diagnosis
 from .obd_data import NewOBDData, OBDData, OBDDataUpdate
@@ -30,50 +47,53 @@ class Status(str, Enum):
 class NewCase(BaseModel):
     """Schema for new cases added via the api."""
 
-    class Config:
-        schema_extra = {
+    model_config = ConfigDict(
+        json_schema_extra={
             "example": {
                 "vehicle_vin": "VIN42",
                 "occasion": Occasion.service_routine,
                 "milage": 42
             }
         }
+    )
 
     vehicle_vin: str
     customer_id: Optional[PydanticObjectId] = None
     occasion: Occasion = Occasion.unknown
-    milage: int = None
+    milage: Optional[int] = None
 
 
 class CaseUpdate(BaseModel):
     """Metadata of a case that can be updated after creation."""
-    timestamp: datetime = None
-    occasion: Occasion = None
-    milage: int = None
-    status: Status = None
+    timestamp: Optional[datetime] = None
+    occasion: Optional[Occasion] = None
+    milage: Optional[int] = None
+    status: Optional[Status] = None
 
 
 class Case(Document):
     """Complete case schema and major db interfacing class."""
 
-    class Config:
-        validate_assignment = True
-        fields = {"schema_version": {"exclude": True}}
+    model_config = ConfigDict(
+        validate_assignment=True
+    )
 
     class Settings:
         name = "cases"
 
     # case descriptions
-    timestamp: datetime = Field(default_factory=datetime.utcnow)
+    timestamp: datetime = Field(default_factory=lambda: datetime.now(UTC))
     occasion: Occasion = Occasion.unknown
-    milage: int = None
+    milage: Optional[int] = None
     status: Status = Status.open
 
     # foreign keys
-    customer_id: Optional[Indexed(PydanticObjectId, unique=False)]
-    vehicle_vin: Indexed(str, unique=False)
-    workshop_id: Indexed(str, unique=False)
-    diagnosis_id: Optional[Indexed(PydanticObjectId)]
+    customer_id: Optional[
+        Annotated[PydanticObjectId, Indexed(unique=False)]
+    ] = None
+    vehicle_vin: Annotated[str, Indexed(unique=False)]
+    workshop_id: Annotated[str, Indexed(unique=False)]
+    diagnosis_id: Optional[Annotated[PydanticObjectId, Indexed()]] = None
 
     # diagnostic data
     timeseries_data: List[TimeseriesData] = []
@@ -85,7 +105,10 @@ class Case(Document):
     obd_data_added: NonNegativeInt = 0
     symptoms_added: NonNegativeInt = 0
 
-    schema_version: int = 0
+    schema_version: int = Field(
+        default=0,
+        exclude=True
+    )
 
     @before_event(Insert)
     async def insert_vehicle(self):
@@ -101,10 +124,10 @@ class Case(Document):
     @classmethod
     async def find_in_hub(
             cls,
-            customer_id: str = None,
-            vin: str = None,
-            workshop_id: str = None
-    ):
+            customer_id: Optional[str] = None,
+            vin: Optional[str] = None,
+            workshop_id: Optional[str] = None
+    ) -> List[Self]:
         """
         Get list of all cases filtered by customer_id, vehicle_vin and
         workshop_id.
@@ -120,7 +143,7 @@ class Case(Document):
         cases = await cls.find(filter).to_list()
         return cases
 
-    async def add_timeseries_data(self, new_data: NewTimeseriesData) -> "Case":
+    async def add_timeseries_data(self, new_data: NewTimeseriesData) -> Self:
         # signal data is stored and converted to ref
         timeseries_data = await new_data.to_timeseries_data()
         # append to data array
@@ -130,15 +153,17 @@ class Case(Document):
         await self.save()
         return self
 
-    async def add_obd_data(self, new_obd_data: NewOBDData) -> "Case":
-        obd_data = OBDData(data_id=self.obd_data_added, **new_obd_data.dict())
+    async def add_obd_data(self, new_obd_data: NewOBDData) -> Self:
+        obd_data = OBDData(data_id=self.obd_data_added,
+                           **new_obd_data.model_dump())
         self.obd_data.append(obd_data)
         self.obd_data_added += 1
         await self.save()
         return self
 
-    async def add_symptom(self, new_symptom: NewSymptom) -> "Case":
-        symptom = Symptom(data_id=self.symptoms_added, **new_symptom.dict())
+    async def add_symptom(self, new_symptom: NewSymptom) -> Self:
+        symptom = Symptom(data_id=self.symptoms_added,
+                          **new_symptom.model_dump())
         self.symptoms.append(symptom)
         self.symptoms_added += 1
         await self.save()
@@ -147,26 +172,28 @@ class Case(Document):
     @staticmethod
     def find_data_in_array(
             data_array: list, data_id: NonNegativeInt
-    ) -> (NonNegativeInt, Union[TimeseriesData, OBDData, Symptom]):
+    ) -> Tuple[NonNegativeInt, Any] | Tuple[None, None]:
         for i, d in enumerate(data_array):
             if d.data_id == data_id:
                 return i, d
 
         return None, None
 
-    def get_timeseries_data(self, data_id: NonNegativeInt) -> TimeseriesData:
+    def get_timeseries_data(
+            self, data_id: NonNegativeInt
+    ) -> TimeseriesData | None:
         _, timeseries_data = self.find_data_in_array(
             data_array=self.timeseries_data, data_id=data_id
         )
         return timeseries_data
 
-    def get_obd_data(self, data_id: NonNegativeInt) -> OBDData:
+    def get_obd_data(self, data_id: NonNegativeInt) -> OBDData | None:
         _, obd_data = self.find_data_in_array(
             data_array=self.obd_data, data_id=data_id
         )
         return obd_data
 
-    def get_symptom(self, data_id: NonNegativeInt) -> Symptom:
+    def get_symptom(self, data_id: NonNegativeInt) -> Symptom | None:
         _, symptom = self.find_data_in_array(
             data_array=self.symptoms, data_id=data_id
         )
@@ -176,7 +203,7 @@ class Case(Document):
         idx, timeseries_data = self.find_data_in_array(
             data_array=self.timeseries_data, data_id=data_id
         )
-        if timeseries_data is not None:
+        if timeseries_data is not None and idx is not None:
             await timeseries_data.delete_signal()
             self.timeseries_data.pop(idx)
             await self.save()
@@ -199,13 +226,13 @@ class Case(Document):
 
     async def update_timeseries_data(
             self, data_id: NonNegativeInt, update: TimeseriesDataUpdate
-    ) -> TimeseriesData:
+    ) -> TimeseriesData | None:
         idx, timeseries_data = self.find_data_in_array(
             data_array=self.timeseries_data, data_id=data_id
         )
-        if timeseries_data is not None:
-            timeseries_data = timeseries_data.dict()
-            timeseries_data.update(update.dict(exclude_unset=True))
+        if timeseries_data is not None and idx is not None:
+            timeseries_data = timeseries_data.model_dump()
+            timeseries_data.update(update.model_dump(exclude_unset=True))
             timeseries_data = TimeseriesData(**timeseries_data)
             self.timeseries_data[idx] = timeseries_data
             await self.save()
@@ -213,13 +240,13 @@ class Case(Document):
 
     async def update_obd_data(
             self, data_id: NonNegativeInt, update: OBDDataUpdate
-    ) -> OBDData:
+    ) -> OBDData | None:
         idx, obd_data = self.find_data_in_array(
             data_array=self.obd_data, data_id=data_id
         )
-        if obd_data is not None:
-            obd_data = obd_data.dict()
-            obd_data.update(update.dict(exclude_unset=True))
+        if obd_data is not None and idx is not None:
+            obd_data = obd_data.model_dump()
+            obd_data.update(update.model_dump(exclude_unset=True))
             obd_data = OBDData(**obd_data)
             self.obd_data[idx] = obd_data
             await self.save()
@@ -227,13 +254,13 @@ class Case(Document):
 
     async def update_symptom(
             self, data_id: NonNegativeInt, update: SymptomUpdate
-    ) -> Symptom:
+    ) -> Symptom | None:
         idx, symptom = self.find_data_in_array(
             data_array=self.symptoms, data_id=data_id
         )
-        if symptom is not None:
-            symptom = symptom.dict()
-            symptom.update(update.dict(exclude_unset=True))
+        if symptom is not None and idx is not None:
+            symptom = symptom.model_dump()
+            symptom.update(update.model_dump(exclude_unset=True))
             symptom = Symptom(**symptom)
             self.symptoms[idx] = symptom
             await self.save()
@@ -270,4 +297,5 @@ class Case(Document):
             # delete via instance to make sure Diagnosis event handlers
             # are also executed
             diag = await Diagnosis.get(self.diagnosis_id)
-            await diag.delete()
+            if diag is not None:
+                await diag.delete()
